refactor(mon-pix): move modulix passage controller to component
