+++ mon-pix/app/templates/module/get.hbs
@@ -1,8 +1,3 @@
 <div class="modulix">
-  <Module::Passage
-    @module={{@model.module}}
-    @retryElement={{this.trackRetry}}
-    @submitAnswer={{this.submitAnswer}}
-    @passage={{@model.passage}}
-  />
+  <Module::Passage @module={{@model.module}} @passage={{@model.passage}} />
 </div>
